requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
enum-compat==0.0.3
idna==3.10
intuit-oauth==1.2.6
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-quickbooks==0.9.12
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
